Fix core dump when tags file pattern has a trailing '\' ... due to size_t being unsigned and a loop that checks for > 0 which can never happen.

+++ ex/ex_subst.c
@@ -1151,7 +1151,7 @@ re_conv(SCR *sp, CHAR_T **ptrnp, size_t *plenp, int *replacedp)
 static int
 re_tag_conv(SCR *sp, CHAR_T **ptrnp, size_t *plenp, int *replacedp)
 {
-	size_t blen, len;
+	int blen, len;
 	int lastdollar;
 	CHAR_T *bp, *p, *t;
